infra: rename test folder
