+++ test/unit_tests/test_braket_quantum_task.py
@@ -16,7 +16,7 @@ from qiskit_braket_provider import (
     BraketQuantumTask,
 )
 from qiskit_braket_provider.providers.braket_quantum_task import retry_if_result_none
-from tests.providers.mocks import MOCK_LOCAL_QUANTUM_TASK, MOCK_PROGRAM_SET_QUANTUM_TASK
+from test.unit_tests.mocks import MOCK_LOCAL_QUANTUM_TASK, MOCK_PROGRAM_SET_QUANTUM_TASK
 
 
 class TestBraketQuantumTask(TestCase):
